runner/requirements.txt
numpy
tensorflow-cpu
boto3
aws-lambda-powertools
